API

API

ScatteredInterpolation.evaluate

ScatteredInterpolation.interpolate

ScatteredInterpolation.Gaussian

ScatteredInterpolation.InverseMultiquadratic

ScatteredInterpolation.InverseQuadratic

ScatteredInterpolation.Multiquadratic

ScatteredInterpolation.NearestNeighbor

ScatteredInterpolation.Polyharmonic

ScatteredInterpolation.Shepard

ScatteredInterpolation.ThinPlate

Functions

ScatteredInterpolation.interpolate — Function.

interpolate(method, points, samples; metric = Euclidean(), returnRBFmatrix = false)

Create an interpolation of the data in samples sampled at the locations defined in points based on the interpolation method method. metric is any of the metrics defined by the Distances package. The RBF matrix used for solving the weights can be returned with the boolean returnRBFmatrix. Note that this option is only valid for RadialBasisFunction interpolations.

points should be an $n×k$ matrix, where $n$ is dimension of the sampled space and $k$ is the number of points. This means that each column in the matrix defines one point.

samples is an $k×m$ array, where $k$ is the number of sampled points (same as for points) and $m$ is the dimension of the sampled data.

The returned ScatteredInterpolant object can be passed to evaluate to interpolate the data to new points.

source

ScatteredInterpolation.evaluate — Function.

evaluate(itp, points)

Evaluate an interpolation object itp at the locations defined in points.

points should be an $n×k$ matrix, where $n$ is dimension of the sampled space and $k$ is the number of points. This means that each column in the matrix defines one point.

source

Types

Radial Basis Functions

ScatteredInterpolation.Multiquadratic — Type.

Multiquadratic(ɛ = 1)

Define a Multiquadratic Radial Basis Function

\[ϕ(r) = \sqrt{1 + (ɛr)^2}\]

source

ScatteredInterpolation.InverseMultiquadratic — Type.

InverseMultiquadratic(ɛ = 1)

Define an Inverse Multiquadratic Radial Basis Function

\[ϕ(r) = \frac{1}{\sqrt{1 + (ɛr)^2}}\]

source

ScatteredInterpolation.Gaussian — Type.

Gaussian(ɛ = 1)

Define a Gaussian Radial Basis Function

\[ϕ(r) = e^{-(ɛr)^2}\]

source

ScatteredInterpolation.InverseQuadratic — Type.

InverseQuadratic(ɛ = 1)

Define an Inverse Quadratic Radial Basis Function

\[ϕ(r) = \frac{1}{1 + (ɛr)^2}\]

source

ScatteredInterpolation.Polyharmonic — Type.

Polyharmonic(k = 1)

Define a Polyharmonic Spline Radial Basis Function

\[ϕ(r) = r^k, k = 1, 3, 5, ... \\ ϕ(r) = r^k ln(r), k = 2, 4, 6, ...\]

source

ScatteredInterpolation.ThinPlate — Type.

ThinPlate()

Define a Thin Plate Spline Radial Basis Function

\[ϕ(r) = r^2 ln(r)\]

This is a shorthand for Polyharmonic(2).

source

Inverse Distance Weighting (Shepard)

ScatteredInterpolation.Shepard — Type.

Shepard(P = 2)

Standard Shepard interpolation with power parameter P.

source

Nearest Neighbor

ScatteredInterpolation.NearestNeighbor — Type.

NearestNeigbor()

Nearest neighbor interpolation.

source